Extract stop() to AbstractPipelineJob
